Bulk upload to github after adding networth calculation.
- Fixed throne page display
- Cleaned up adding a military to province so that’s it’s 1to1
- Province object can be accessed from military and vice versa.

Tons of other small shit. Will do better at committing each change now
the re-structure is finished. Was messy.

+++ newtopia/ntmeta/models.py
@@ -1,7 +1,5 @@
 from django.db import models
 
-from ntgame.models import Race, Province
-
 # Create your models here.
 ''' This is a meta object - potentially move to ntmeta or ntcore (depending on)
 scope of future app. '''
@@ -18,7 +16,7 @@ class Rule(models.Model):
 class NetworthValue(models.Model):
 
     prop = models.CharField(max_length=40)
-    value = models.IntegerField(default=0)
+    value = models.FloatField(default=0)
 
     def __str__(self):
         return self.prop
